refactor: 장소 목록 가져오는 방식 변경 (#32)
