+++ src/components/buttons/LikeButton.jsx
@@ -1,7 +1,7 @@
 import { IoHeartOutline, IoHeart } from 'react-icons/io5';
 import { useToggleLikes } from '../../libs/hooks/useLikes';
 import useAuthStore from '../../stores/useAuthstore';
-import { useNavigate, useSearchParams } from 'react-router-dom';
+import { useSearchParams } from 'react-router-dom';
 
 /**
  * * 장소의 좋아요 기능을 추가한 좋아요 버튼 컴포넌트
@@ -13,7 +13,6 @@ import { useNavigate, useSearchParams } from 'react-router-dom';
  * <LikePlaceButton size={30} placeId={id} isLiked={isLiked} />
  */
 export const LikePlaceButton = ({ isLiked, placeId, ...props }) => {
-  const navigate = useNavigate();
   const [searchParams] = useSearchParams();
   const {
     userInfo: { id: userId },
@@ -25,14 +24,10 @@ export const LikePlaceButton = ({ isLiked, placeId, ...props }) => {
   // 좋아요 버튼 클릭 리스너
   const handleClickLikeButton = (e) => {
     e.stopPropagation();
-    if (!userId) {
-      alert('로그인이 필요합니다!');
-      navigate('/signin');
-    }
     toggleLike({ userId, placeId });
   };
 
-  return <LikeButton isLiked={isLiked} onClick={handleClickLikeButton} {...props} />;
+  return userId && <LikeButton isLiked={isLiked} onClick={handleClickLikeButton} {...props} />;
 };
 
 /**
