+++ src/components/home/HomeCardContainer.jsx
@@ -1,10 +1,20 @@
 import HomeCard from './HomeCard';
 import SideBar from './SideBar';
-import { useNavigate } from 'react-router-dom';
+import { useNavigate, useSearchParams } from 'react-router-dom';
 import homeUtils from '../../libs/utils/homeUtils';
+import useAuthStore from '../../stores/useAuthstore';
+import useGetPlaces from '../../libs/hooks/useGetPlaces';
 
-const HomeCardContainer = ({ places }) => {
+const HomeCardContainer = () => {
   const navigate = useNavigate();
+  const {
+    userInfo: { id: userId },
+  } = useAuthStore();
+  const [searchParams] = useSearchParams();
+
+  // 쿼리스트링에 따라 데이터 다르게 가져오기
+  const category = searchParams.get('category');
+  const { data: places, isLoading } = useGetPlaces(userId, category);
 
   // * 카테고리 정렬 핸들러 함수
   const handleCategory = (e) => {
@@ -12,6 +22,8 @@ const HomeCardContainer = ({ places }) => {
     homeUtils.handleCategoryMove(categoryName, navigate);
     homeUtils.scrollToTop();
   };
+
+  if (isLoading) return <div>loading...</div>;
 
   return (
     <div className="lg:w-full lg:max-w-3xl m-auto flex flex-wrap gap-7 justify-evenly p-4 sm:w-1/2 md:gap-20">
