+++ src/libs/api/placesData.js
@@ -1,5 +1,4 @@
 import homeUtils from '../utils/homeUtils';
-import { fetchLikes } from './likesData';
 import { supabase } from './supabaseClient';
 
 /**
@@ -7,10 +6,14 @@ import { supabase } from './supabaseClient';
  * @param {string} userId - 현재 로그인한 유저의 uuid
  * @param {string} category - 현재 페이지의 category ('Restaurant', 'Cafe' 또는 null)
  * @returns {Object[]} 응답 데이터 배열 (장소 목록)
- *    { address, category1, category2, charge, coordinates, created_at, description, id, image, isLiked, tel, title, url }
+ *    { address, category1, category2, charge, coordinates, created_at, description, id, image, is_liked, tel, title, url }
  */
 const fetchPlacesData = async (userId, category) => {
-  let query = supabase.from('places').select(`*`);
+  let query = userId
+    ? supabase.rpc('get_places_with_likes', {
+        user_id: userId,
+      })
+    : supabase.from('places').select(`*`);
 
   // 카테고리 있을 경우 eq 쿼리 추가
   if (category) {
@@ -20,19 +23,12 @@ const fetchPlacesData = async (userId, category) => {
 
   // 데이터 불러오기
   const { data: places, error } = await query;
-  const likes = await fetchLikes(userId);
 
   if (error) {
     throw error;
   }
 
-  // 좋아요 여부 넣은 데이터로 가공
-  const placesWithLikeStuats = places.map((place) => ({
-    ...place,
-    isLiked: likes.some((likePlace) => likePlace.place_id === place.id),
-  }));
-
-  return placesWithLikeStuats;
+  return places;
 };
 
 export default fetchPlacesData;
